move extra parameter validation to own file
