+++ lib/settings.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var settings = settingsWithDefaults();
+var settings;
 
 /**
  * Retrieves the default hippie-swagger settings, honoring overrides
